Simulation of Random MAC Transmission Protocol using NS2
Simulation of a wireless sensor network consisting of 100 source nodes
and 1 sink node. Network and Protocol specifications were given in the
form of a TCL script. Protocol changes were made by extending original
NS2 code in C. Developed a python script to read the trace file and to
calculate the average packet delivery probability of the network.